// src/app/api/generated/model/create-invitacion-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Datos de la invitacion a emitir
 */
export interface CreateInvitacionRequest { 
    /**
     * Sede del vinculo propuesto. Ausente o null para alcance organizacion
     */
    consultorioId?: number;
    /**
     * Email de la persona a invitar. Puede no tener cuenta todavia
     */
    email: string;
    /**
     * Rol con el que quedaria vinculada al aceptar
     */
    roleCode: string;
}